+++ tests/backend/dep/test_core.py
@@ -28,6 +28,16 @@ def test_version_unmet(platform):
         assert not dependencies_in_sync([Requirement('binary>9000')], venv.sys_path)
 
 
+def test_marker_met(platform):
+    with TempVirtualEnv(sys.executable, platform) as venv:
+        assert dependencies_in_sync([Requirement('binary; python_version < "1"')], venv.sys_path)
+
+
+def test_marker_unmet(platform):
+    with TempVirtualEnv(sys.executable, platform) as venv:
+        assert not dependencies_in_sync([Requirement('binary; python_version > "1"')], venv.sys_path)
+
+
 def test_extra_no_dependencies(platform):
     with TempVirtualEnv(sys.executable, platform) as venv:
         platform.run_command(['pip', 'install', 'binary'], check=True, capture_output=True)