+++ docs/meta/history.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ***Added:***
 
 - Add ability to set the non-preview variant of versions
+- Improve collection of information about Python interpreters within virtual environments
 - Bump the minimum supported version of Hatchling
 
 ### [1.0.0rc8](https://github.com/ofek/hatch/releases/tag/hatch-v1.0.0rc8) - 2022-01-08 ### {: #hatch-v1.0.0rc8 }