+++ docs/meta/history.md
@@ -16,6 +16,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Improve collection of information about Python interpreters within virtual environments
 - Bump the minimum supported version of Hatchling
 
+***Fixed:***
+
+- Relax restriction on the contents of `build-system.requires`
+
 ### [1.0.0rc8](https://github.com/ofek/hatch/releases/tag/hatch-v1.0.0rc8) - 2022-01-08 ### {: #hatch-v1.0.0rc8 }
 
 ***Added:***