+++ docs/meta/history.md
@@ -20,6 +20,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - Relax restriction on the contents of `build-system.requires`
 - The `version` command now properly handles statically defined versions
+- Fix typo preventing the Python executable fallback for environments when there is no `python` along `PATH`
 
 ### [1.0.0rc8](https://github.com/ofek/hatch/releases/tag/hatch-v1.0.0rc8) - 2022-01-08 ### {: #hatch-v1.0.0rc8 }